+++ basic_autonomy/CMakeLists.txt
@@ -1,5 +1,5 @@
-#
-# Copyright (C) 2021 LEIDOS.
+
+# Copyright (C) 2022 LEIDOS.
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may not
 # use this file except in compliance with the License. You may obtain a copy of
@@ -12,110 +12,56 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations under
 # the License.
-#
 
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.5)
 project(basic_autonomy)
 
+# Declare carma package and check ROS version
 find_package(carma_cmake_common REQUIRED)
-carma_check_ros_version(1)
+carma_check_ros_version(2)
 carma_package()
 
-set( CATKIN_DEPS
-  roscpp
-  std_msgs
-  cav_msgs
-  cav_srvs
-  carma_utils
-  trajectory_utils
-  autoware_msgs
-  carma_wm
-  tf
-  tf2
-  tf2_geometry_msgs
-  roslib
-  carma_debug_msgs
-)
+## Find dependencies using ament auto
+find_package(ament_cmake_auto REQUIRED)
+ament_auto_find_build_dependencies()
 
-## Find catkin macros and libraries
-find_package(catkin REQUIRED COMPONENTS
- ${CATKIN_DEPS}
-)
+# Name build targets
+set(node_lib basic_autonomy_lib)
 
-## System dependencies are found with CMake's conventions
-find_package(Boost)
-find_package(Eigen3 REQUIRED)
-
-###################################
-## catkin specific configuration ##
-###################################
-
-catkin_package(
-  INCLUDE_DIRS include
-  LIBRARIES ${PROJECT_NAME}
-  CATKIN_DEPENDS ${CATKIN_DEPS}
-  DEPENDS Boost EIGEN3
-)
-
-###########
-## Build ##
-###########
-
+# Includes
 include_directories(
   include
-  ${catkin_INCLUDE_DIRS}
-  ${Boost_INCLUDE_DIRS}
-  ${EIGEN3_INCLUDE_DIRS}
 )
 
-add_library(
-  ${PROJECT_NAME} 
-  src/basic_autonomy.cpp
-  src/smoothing/BSpline.cpp
-  src/smoothing/filters.cpp
-  src/log/log.cpp
-  src/helper_functions.cpp
+# Build
+ament_auto_add_library(${node_lib} SHARED
+        src/basic_autonomy.cpp
+        src/helper_functions.cpp
+        src/log/log.cpp
+        src/smoothing/BSpline.cpp
+        src/smoothing/filters.cpp
 )
 
-add_dependencies(${PROJECT_NAME}
-  ${${PROJECT_NAME}_EXPORTED_TARGETS}
-  ${catkin_EXPORTED_TARGETS}
-)
 target_link_libraries(
-  ${PROJECT_NAME}
-  ${catkin_LIBRARIES}
-  ${Boost_LIBRARIES}
-)
-
-#target_link_libraries(basic_autonomy ${catkin_LIBRARIES} ${Boost_LIBRARIES})
-#add_dependencies(basic_autonomy ${catkin_EXPORTED_TARGETS} ${${PROJECT_NAME}_EXPORTED_TARGETS})
-
-
-#############
-## Install ##
-#############
-
-## Install Other Resources
-
-
-## Install C++
-install(TARGETS ${PROJECT_NAME} basic_autonomy
-  ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
-
-install(DIRECTORY include/${PROJECT_NAME}/
-  DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+        ${node_lib}
 )
 
 
-#############
-## Testing ##
-#############
-                           
-catkin_add_gmock(${PROJECT_NAME}-test 
-  test/test_waypoint_generation.cpp
-)
-target_link_libraries(${PROJECT_NAME}-test basic_autonomy ${catkin_LIBRARIES})
+# Testing
+if(BUILD_TESTING)  
 
+  find_package(ament_lint_auto REQUIRED)
+  ament_lint_auto_find_test_dependencies() # This populates the ${${PROJECT_NAME}_FOUND_TEST_DEPENDS} variable
+
+  ament_add_gtest(test_basic_autonomy test/test_waypoint_generation.cpp)
+
+  ament_target_dependencies(test_basic_autonomy ${${PROJECT_NAME}_FOUND_TEST_DEPENDS})
+
+  target_link_libraries(test_basic_autonomy ${node_lib})
+
+endif()
+
+# Install
+ament_auto_package(
+        INSTALL_TO_SHARE resource
+)
