+++ basic_autonomy/README.md
@@ -1,0 +1,3 @@
+# basic_autonomy
+
+The basic_autonomy library includes functions for waypoint processing, which are used by tactical plugins in order to generate detailed trajectories for the CARMA system.
