+++ basic_autonomy/include/basic_autonomy_ros2/helper_functions.hpp
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-#include <basic_autonomy_ros2/basic_autonomy.hpp>
+#include <basic_autonomy/basic_autonomy.hpp>
 
 namespace basic_autonomy
 {
