+++ basic_autonomy/include/basic_autonomy_ros2/smoothing/BSpline.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 /*
- * Copyright (C) 2019-2021 LEIDOS.
+ * Copyright (C) 2019-2022 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -17,8 +17,8 @@
  */
 
 #include <vector>
-#include <carma_wm/Geometry.h>
-#include <basic_autonomy/smoothing/SplineI.h>
+#include <carma_wm_ros2/Geometry.hpp>
+#include <basic_autonomy/smoothing/SplineI.hpp>
 #include <unsupported/Eigen/Splines>
 
 namespace basic_autonomy
