+++ basic_autonomy/package.xml
@@ -1,14 +1,11 @@
 <?xml version="1.0"?>
 
 <!--  
- Copyright (C) 2021 LEIDOS.
-
+ Copyright (C) 2021-2022 LEIDOS.
  Licensed under the Apache License, Version 2.0 (the "License"); you may not
  use this file except in compliance with the License. You may obtain a copy of
  the License at
-
  http://www.apache.org/licenses/LICENSE-2.0
-
  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -18,24 +15,33 @@
 
 <package format="3">
   <name>basic_autonomy</name>
-  <version>3.2.0</version>
-  <description>The Basic Autonomy Library includes logic for waypoint generation in CARMA Platform.</description>
-  <maintainer email="CARMA@dot.gov">carma</maintainer>
-  <license>Apache License 2.0</license>
-  <author email="CARMA@dot.gov">carma</author>
-  <buildtool_depend>catkin</buildtool_depend>
-  <depend>roscpp</depend>
-  <depend>std_msgs</depend>
-  <depend>cav_srvs</depend>
-  <depend>cav_msgs</depend>
-  <depend>carma_utils</depend>
-  <depend>autoware_msgs</depend>
-  <depend>carma_wm</depend>
-  <depend>tf</depend>
-  <depend>tf2</depend>
-  <depend>tf2_geometry_msgs</depend> 
-  <depend>trajectory_utils</depend>   
-  <depend>roslib</depend>
-  <depend>carma_debug_msgs</depend> 
+  <version>4.0.0</version>
+  <description>The basic_autonomy package</description>
+
+  <maintainer email="carma@dot.gov">carma</maintainer>
+
+  <license>Apache 2.0</license>
+  
+  <buildtool_depend>ament_cmake</buildtool_depend>
   <build_depend>carma_cmake_common</build_depend>
+  <build_depend>ament_auto_cmake</build_depend>
+
+  <depend>rclcpp</depend>
+  <depend>carma_ros2_utils</depend>
+  <depend>rclcpp_components</depend>
+  <depend>carma_planning_msgs</depend>
+  <depend>carma_debug_ros2_msgs</depend>
+  <depend>lanelet2_core</depend>
+  <depend>trajectory_utils_ros2</depend>
+  <depend>carma_wm_ros2</depend>
+  <depend>autoware_msgs</depend>
+  <depend>tf2</depend>
+  <depend>autoware_auto_msgs</depend>
+
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_cmake_gtest</test_depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
 </package>
