+++ basic_autonomy/src/smoothing/filters.cpp
@@ -1,6 +1,5 @@
-
 /*
- * Copyright (C) 2021 LEIDOS.
+ * Copyright (C) 2021-2022 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -15,7 +14,8 @@
  * the License.
  */
 
-#include <basic_autonomy/smoothing/filters.h>
+#include <basic_autonomy/smoothing/filters.hpp>
+
 namespace basic_autonomy
 {
 namespace smoothing
