--- basic_autonomy_ros2/README.md
@@ -1,3 +1,0 @@
-# basic_autonomy_ros2
-
-The basic_autonomy_ros2 library includes functions for waypoint processing, which are used by tactical plugins in order to generate detailed trajectories for the CARMA system.
