+++ cooperative_lanechange/test/test_cooperative_plugin.cpp
@@ -80,7 +80,7 @@ namespace cooperative_lanechange
     TEST(CooperativeLaneChangePlugin,TestLaneChangefunctions)
     {
         // File to process. 
-        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy_ros2");
+        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy");
         std::string file = "/resource/map/town01_vector_map_lane_change.osm";
         file = path.append(file);
 
@@ -207,7 +207,7 @@ namespace cooperative_lanechange
 
     TEST(CooperativeLaneChangePlugin,Testcurrentgapcb){
         // File to process. 
-        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy_ros2");
+        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy");
         std::string file = "/resource/map/town01_vector_map_lane_change.osm";        
         file = path.append(file);
 
@@ -335,7 +335,7 @@ namespace cooperative_lanechange
         //Tests behavior when there is no path from roadway object to subject vehicle
 
         // File to process. 
-        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy_ros2");
+        std::string path = ament_index_cpp::get_package_share_directory("basic_autonomy");
         std::string file = "/resource/map/town01_vector_map_lane_change.osm"; 
         file = path.append(file);
         lanelet::Id start_id = 107;
