+++ inlanecruising_plugin/include/inlanecruising_plugin/inlanecruising_plugin.hpp
@@ -24,7 +24,7 @@
 #include <carma_ros2_utils/carma_lifecycle_node.hpp>
 #include <boost/geometry.hpp>
 #include <carma_wm_ros2/Geometry.hpp>
-#include <basic_autonomy_ros2/basic_autonomy.hpp>
+#include <basic_autonomy/basic_autonomy.hpp>
 #include <carma_planning_msgs/srv/plan_trajectory.hpp>
 #include <carma_wm_ros2/WMListener.hpp>
 #include <functional>
