+++ platooning_tactical_plugin/package.xml
@@ -37,7 +37,7 @@
   <depend>tf2</depend>
   <depend>tf2_geometry_msgs</depend>
   <depend>trajectory_utils_ros2</depend>
-  <depend>basic_autonomy_ros2</depend>
+  <depend>basic_autonomy</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_cmake_gtest</test_depend>
