+++ pure_pursuit_wrapper/include/pure_pursuit_wrapper/pure_pursuit_wrapper.hpp
@@ -26,7 +26,7 @@
 #include <carma_guidance_plugins/control_plugin.hpp>
 #include <pure_pursuit/pure_pursuit.hpp>
 #include <gtest/gtest_prod.h>
-#include <basic_autonomy_ros2/basic_autonomy.hpp>
+#include <basic_autonomy/basic_autonomy.hpp>
 
 namespace pure_pursuit_wrapper {
