+++ pure_pursuit_wrapper/package.xml
@@ -34,7 +34,7 @@
   <depend>carma_guidance_plugins</depend>
   <depend>pure_pursuit</depend>
   <depend>motion_testing</depend>
-  <depend>basic_autonomy_ros2</depend>
+  <depend>basic_autonomy</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_cmake_gtest</test_depend>
